pyproject.toml
[project]
name = "raman-id"
version = "0.1.0"
dependencies = [
    "python-dotenv",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"